Move Stream and related SDK modules to messaging directory
Previously, Stream, Future, and related messaging modules, used
by both clients and transaction processors, was erroneously
includes in the `/client` directory. They have been moved to
a new `/messaging` directory, and State modules have been moved
to `/processor`, to match the Python SDK.
