+++ consensus/poet/families/sawtooth_validator_registry/validator_registry/processor/handler.py
@@ -25,7 +25,7 @@ from cryptography.hazmat.primitives.asymmetric import padding
 from cryptography.exceptions import InvalidSignature
 
 from sawtooth_sdk.processor.state import StateEntry
-from sawtooth_sdk.client.future import FutureTimeoutError
+from sawtooth_sdk.messaging.future import FutureTimeoutError
 from sawtooth_sdk.processor.exceptions import InvalidTransaction
 from sawtooth_sdk.processor.exceptions import InternalError
 from sawtooth_sdk.protobuf.transaction_pb2 import TransactionHeader
