+++ core_transactions/config/sawtooth_config/processor/handler.py
@@ -18,7 +18,7 @@ import hashlib
 import base64
 
 from sawtooth_sdk.processor.state import StateEntry
-from sawtooth_sdk.client.future import FutureTimeoutError
+from sawtooth_sdk.messaging.future import FutureTimeoutError
 from sawtooth_sdk.processor.exceptions import InvalidTransaction
 from sawtooth_sdk.processor.exceptions import InternalError
 from sawtooth_sdk.protobuf.transaction_pb2 import TransactionHeader
