+++ sdk/java/src/main/java/sawtooth/sdk/messaging/SendReceiveThread.java
@@ -12,7 +12,7 @@
  limitations under the License.
 ------------------------------------------------------------------------------*/
 
-package sawtooth.sdk.client;
+package sawtooth.sdk.messaging;
 
 import com.google.protobuf.InvalidProtocolBufferException;
