+++ sdk/java/src/main/java/sawtooth/sdk/messaging/Stream.java
@@ -12,7 +12,7 @@
  limitations under the License.
 ------------------------------------------------------------------------------*/
 
-package sawtooth.sdk.client;
+package sawtooth.sdk.messaging;
 
 import com.google.protobuf.ByteString;
