+++ sdk/javascript/messaging/deferred.js
