+++ sdk/javascript/messaging/stream.js
