+++ sdk/javascript/processor/state.js
