+++ sdk/python/sawtooth_sdk/__init__.py
@@ -16,5 +16,6 @@
 
 __all__ = [
     'client',
+    'messaging',
     'processor'
 ]
