+++ sdk/python/sawtooth_sdk/messaging/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2017 Intel Corporation
+# Copyright 2016 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,6 +15,7 @@
 
 
 __all__ = [
-    'config',
-    'log'
+    'exceptions',
+    'future',
+    'stream'
 ]
