+++ sdk/python/sawtooth_sdk/messaging/exceptions.py
